requirements.txt
langchain
langchain_core
python-dotenv
streamlit
langchain_community
langchain-huggingface

fastapi
langserve
uvicorn
sse_starlette